requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
pillow==11.3.0
Pygments==2.19.2
sqlparse==0.5.3
django-filter
